feat: Rename token file to index according to standard
