+++ src/modules/roles/roles.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Role } from './entities/role.entity';
 import { RoleController } from './controllers/role.controller';
 import { RoleService } from './services/role.service';
-import { ROLE_REPOSITORY } from './tokens/tokens';
+import { ROLE_REPOSITORY } from './tokens';
 import { TypeormRolesRepository } from './repositories/typeorm-roles.repository';
 
 @Module({
